refactor: Move all kubejs related classes into the kubejs package
