+++ common/src/main/java/net/impleri/playerskills/integration/kubejs/PlayerSkillsPlugin.java
@@ -1,5 +1,6 @@
 package net.impleri.playerskills.integration.kubejs;
 
+import dev.architectury.event.events.common.LifecycleEvent;
 import dev.latvian.mods.kubejs.KubeJSPlugin;
 import dev.latvian.mods.kubejs.script.BindingsEvent;
 import dev.latvian.mods.kubejs.util.AttachedData;
@@ -16,22 +17,25 @@ import net.impleri.playerskills.variant.basic.BasicSkillType;
 import net.impleri.playerskills.variant.numeric.NumericSkillType;
 import net.impleri.playerskills.variant.specialized.SpecializedSkillType;
 import net.impleri.playerskills.variant.tiered.TieredSkillType;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.entity.player.Player;
 
 public class PlayerSkillsPlugin extends KubeJSPlugin {
-    public static void modifySkills() {
-        EventsBinding.MODIFICATION.post(new SkillsModificationEventJS(Registries.SKILLS.types));
-    }
-
     private final PlayerSkillsKubeJSWrapper skillWrapper = new PlayerSkillsKubeJSWrapper();
 
     @Override
     public void init() {
+        LifecycleEvent.SERVER_STARTING.register(this::onServerStart);
         SkillChangedEvent.EVENT.register(this::onSkillChange);
         Registries.SKILLS.addType(BasicSkillType.name.toString(), BasicSkillJS.Builder.class, BasicSkillJS.Builder::new);
         Registries.SKILLS.addType(NumericSkillType.name.toString(), NumericSkillJS.Builder.class, NumericSkillJS.Builder::new);
         Registries.SKILLS.addType(TieredSkillType.name.toString(), TieredSkillJS.Builder.class, TieredSkillJS.Builder::new);
         Registries.SKILLS.addType(SpecializedSkillType.name.toString(), SpecializedSkillJS.Builder.class, SpecializedSkillJS.Builder::new);
+    }
+
+    private void onServerStart(MinecraftServer minecraftServer) {
+        // Trigger skills modification event
+        EventsBinding.MODIFICATION.post(new SkillsModificationEventJS(Registries.SKILLS.types));
     }
 
     @Override
